# frontend/public/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anuvaad API",
    "description": "**Anuvaad** \u2014 AI-powered code translation platform.\n\nTranslate code to plain English, English to code, and between 35+ languages.\n\nAll new clients should use the `/api/v1/` prefix. Legacy `/api/` routes are deprecated and will be removed on **2027-01-01**.",
    "contact": {
      "name": "Anuvaad Support",
      "url": "http://localhost:3000/"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/sso/tenants/{tenant_domain}": {
      "get": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Get Tenant Sso Status",
        "description": "Check whether a given corporate domain is enrolled in Enterprise SSO.",
        "operationId": "get_tenant_sso_status_api_v1_sso_tenants__tenant_domain__get",
        "parameters": [
          {
            "name": "tenant_domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sso/saml/authorize": {
      "post": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Saml Authorize",
        "description": "Initiate SP-initiated SAML flow for an enterprise corporate email.",
        "operationId": "saml_authorize_api_v1_sso_saml_authorize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SAMLAuthorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sso/saml/callback": {
      "post": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Saml Callback",
        "description": "Assertion Consumer Service (ACS) endpoint to validate SAML response.\n\nParses SAML assertion, verifies signature, and issues authenticated enterprise session.",
        "operationId": "saml_callback_api_v1_sso_saml_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SAMLCallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sso/tenants/{tenant_domain}": {
      "get": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Get Tenant Sso Status",
        "description": "Check whether a given corporate domain is enrolled in Enterprise SSO.",
        "operationId": "get_tenant_sso_status_api_sso_tenants__tenant_domain__get",
        "parameters": [
          {
            "name": "tenant_domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sso/saml/authorize": {
      "post": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Saml Authorize",
        "description": "Initiate SP-initiated SAML flow for an enterprise corporate email.",
        "operationId": "saml_authorize_api_sso_saml_authorize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SAMLAuthorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sso/saml/callback": {
      "post": {
        "tags": [
          "Enterprise SSO"
        ],
        "summary": "Saml Callback",
        "description": "Assertion Consumer Service (ACS) endpoint to validate SAML response.\n\nParses SAML assertion, verifies signature, and issues authenticated enterprise session.",
        "operationId": "saml_callback_api_sso_saml_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SAMLCallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SAMLAuthorizeRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "redirect_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect Url"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SAMLAuthorizeRequest"
      },
      "SAMLCallbackRequest": {
        "properties": {
          "SAMLResponse": {
            "type": "string",
            "title": "Samlresponse"
          },
          "RelayState": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relaystate"
          }
        },
        "type": "object",
        "required": [
          "SAMLResponse"
        ],
        "title": "SAMLCallbackRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}